% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structurednessindices.R
\name{c_manifoldness}
\alias{c_manifoldness}
\title{c-manifoldness
calculates c-manifoldness as the maximal correlation coefficient}
\usage{
c_manifoldness(confs)
}
\arguments{
\item{confs}{a numeric matrix or data frame with two columns}
}
\description{
c-manifoldness
calculates c-manifoldness as the maximal correlation coefficient
}
\examples{
x<--100:100
y<-sqrt(100^2-x^2)
confs<-cbind(x,y)
c_manifoldness(confs)
}
