% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structurednessindices.R
\name{c_regularity}
\alias{c_regularity}
\title{c-regularity 
calculates c-regularity as 1 - OPTICS cordillera for k=2. The higher the more regular.}
\usage{
c_regularity(confs, q = 2, epsilon = 2 * max(dist(confs)),
  distmeth = "euclidean", dmax = NULL, digits = 10, scale = 4, ...)
}
\arguments{
\item{confs}{a numeric matrix or a dist object}

\item{q}{The norm used for the Cordillera. Defaults to 2.}

\item{epsilon}{The epsilon parameter for OPTICS (called epsilon_max in the paper). Defaults to 2 times the maximum distance between any two points.}

\item{distmeth}{The distance to be computed if X is not a symmetric matrix or a dist object (otherwise ignored). Defaults to Euclidean distance.}

\item{dmax}{The winsorization value for the highest allowed reachability. If used for comparisons this should be supplied. If no value is supplied, it is NULL (default), then dmax is taken from the data as minimum of epsilon or the largest reachability.}

\item{digits}{The precision to round the raw Cordillera and the norm factor. Defaults to 10.}

\item{scale}{Should X be scaled if it is an asymmetric matrix or data frame? Can take values TRUE or FALSE or a numeric value. If TRUE or 1, standardisation is to mean=0 and sd=1. If 2, no centering is applied and scaling of each column is done with the root mean square of each column. If 3, no centering is applied and scaling of all columns is done as X/max(standard deviation(allcolumns)). If 4, no centering is applied and scaling of all columns is done as X/max(rmsq(allcolumns)). If FALSE, 0 or any other numeric value, no standardisation is applied. Defaults to 4.}

\item{...}{Additional arguments to be passed to \code{\link[dbscan:optics]{optics}}}
}
\description{
c-regularity 
calculates c-regularity as 1 - OPTICS cordillera for k=2. The higher the more regular.
}
\examples{
hpts <- sp:::genHexGrid(dx=0.9, ll=c(-2, -2), ur=c(2, 2))
plot(hpts[,1],hpts[,2],pch=19,asp=1)
c_regularity(hpts)
}
