% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structurednessindices.R
\name{c_shepardness}
\alias{c_shepardness}
\title{c-shepardness 
calculates the c-shepardness as the correlation between a loess smoother of the transformed distances and the transformed dissimilarities}
\usage{
c_shepardness(object)
}
\arguments{
\item{object}{an object of class smacofP}
}
\value{
a numeric value
}
\description{
c-shepardness 
calculates the c-shepardness as the correlation between a loess smoother of the transformed distances and the transformed dissimilarities
}
\examples{
delts<-smacof::kinshipdelta
res<-smacofx::postmds(delts)
c_shepardness(res)
}
