% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structurednessindices.R
\name{c_skinniness}
\alias{c_skinniness}
\title{c-skinniness}
\usage{
c_skinniness(conf)
}
\arguments{
\item{conf}{A numeric matrix.}
}
\description{
Measures the c-skinniness structure
}
\examples{
delts<-smacof::kinshipdelta
conf<-smacof::smacofSym(delts)$conf
plot(conf,pch=19,asp=1)
c_skinniness(conf)
}
