% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structurednessindices.R
\name{c_striatedness}
\alias{c_striatedness}
\title{c-striatedness}
\usage{
c_striatedness(conf)
}
\arguments{
\item{conf}{A numeric matrix.}
}
\description{
Measures the c-striatedness structure
}
\examples{
delts<-smacof::kinshipdelta
conf<-smacof::smacofSym(delts)$conf
plot(conf,pch=19,asp=1)
c_striatedness(conf)
}
