% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structurednessindices.R
\name{c_stringiness}
\alias{c_stringiness}
\title{c-stringiness}
\usage{
c_stringiness(conf)
}
\arguments{
\item{conf}{A numeric matrix.}
}
\description{
Measures the c-stringiness structure
}
\examples{
delts<-smacof::kinshipdelta
conf<-smacof::smacofSym(delts)$conf
plot(conf,pch=19,asp=1)
c_stringiness(conf)
}
