% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAClimateIndicatorsCountyMedian-data.R
\name{CAClimateIndicatorsCountyMedian}
\alias{CAClimateIndicatorsCountyMedian}
\title{Climate Change Indicators of Californian Counties}
\format{
A data frame with 58 rows and 52 variables   
\describe{
\item{county} The county name identifier
\item{vuln_CA} the vulnerability index
 \item{degFA2} County average 95th percentile daily maximum temperature in Fahrenheit from May 1 to September 30 over the historical period (1971-2000) under the climate scenario A2. These are averaged values for 4 different climate models. The source was Table 7 of Cooley et al. (2012).
#'  \item{degFB1} County average 95th percentile daily maximum temperature in Fahrenheit from May 1 to September 30 over the historical period (1971-2000) under the climate scenario B1. These are averaged values for 4 different climate models. The source was Table 7 of Cooley et al. (2012).
\item{heatB1_FYY_TYY} Projected average number of days where the daily maximum temperature exceeds the high-heat threshold (see above) over periods 2010-2039, 2040-2069 and 2070-2099. Projections are based on the B1 scenario and are averaged for four downscaled climate models. The source was Table 7 of Cooley et al. (2012). The variables are heatB1_FYY_TYY, where FYY is the "from year" and TYY the "to year", so heatB1_71_00 is the value of a county under B1 from 1971 to 2000.
\item{heatA2_FYY_TYY} Projected average number of days where the daily maximum temperature exceeds the high-heat threshold (see above) over periods 2010-2039, 2040-2069 and 2070-2099. Projections are based on the A2 scenario and are averaged for four downscaled climate models. The source was Table 7 of Cooley et al. (2012). The variables are heatA2_FYY_TYY, where FYY is the "from year" and TYY the "to year", so heatA2_71_00 is the value of a county under A2 from 1971 to 2000.  
\item{flood_2000} The percentage of a county's census block area vulnerable to unimpeded coastal flooding under baseline conditions (2000). The raw data were obtained from Heberger (2009) . From the census block areas we computed an area-weighted percentage for each county. 
\item{flood_2100} The projected percentage of a county's census block area vulnerable to unimpeded coastal flooding with a 1.4-meter (55-inch) sea-level rise (projected for 2100). The raw data were obtained from Heberger (2009). From the census block areas we computed an area-weighted percentage for each county.
\item{basfA2_YYYY} The median aggregated CCSM3 observed or projected annual baseflow for year YYYY under scenario A2 by county (past years are observed, future years are projected). The source of the raw data was California Energy Commission (2008). The variables are basfA2_YYYY, where and YYYY is any of the years 2000, 2049 and 2099 (observed or projected), so basfA2_2069 is the projected median annual baseflow in that county under A2 in 2069.
\item{basfB1_YYYY} The median aggregated CCSM3 observed or projected annual baseflow for year YYYY under scenario B1 by county (past years are observed, future years are projected). The source of the raw data was California Energy Commission (2008). The variables are basfB1_YYYY, where and YYYY is any of the years 2000, 2049 and 2099 (observed or projected), so basfB1_2069 is the projected median annual baseflow in that county under B1 in 2069.
\item{evapA2_YYYY} The median aggregated Community Climate System Model v.3 (CCSM3) projected annual actual evapotranspiration for year YYYY under scenarios A2 by county. The source of the raw data was California Energy Commission (2008). The variables are evapA2_YYYY where YYYY is any of the years 2000, 2049 and 2099 (observed or projected), so evapA2_2069 is the projected median evaporation in that county under A2 in 2069.
\item{evapB1_YYYY} The median aggregated Community Climate System Model v.3 (CCSM3) projected annual actual evapotranspiration for year YYYY under scenarios B1 by county. The source of the raw data was California Energy Commission (2008). The variables are evapA2_YYYY where YYYY is any of the years 2000, 2049 and 2099 (observed or projected), so evapB1_2069 is the projected median evaporation in that county under B1 in 2069.
\item{prcpA2_YYYY} The median aggregated CCSM3 projected annual precipitation for years YYYY under scenario A2 by county. The source of the raw data was California Energy Commission (2008). The variables are prcpA2_YYYY, where YYYY is any of the years 2000, 2049 and 2099 (observed or projected), so prcpA2_2069 is the projected median percipitation in that county under A2 in 2069.
\item{prcpB1_YYYY} The median aggregated CCSM3 projected annual precipitation for years YYYY under scenario B1 by county. The source of the raw data was California Energy Commission (2008). The variables are prcpB1_YYYY, where YYYY is any of the years 2000, 2049 and 2099 (observed or projected), so prcpA2_2069 is the projected median percipitation in that county under B1 in 2069.
\item{smclA2_YYYY} The median aggregated CCSM3 projected annual fractional moisture in the entire soil column for year YYYY under scenario A2 by county. The source of the raw data was California Energy Commission (2008). The variables are smclA2_YYYY where YYYY is any of the years 2000, 2049 and 2099 (observed or projected), so smclA2_2069 is the projected median soil column moisture in that county under A2 in 2069.
\item{smclB1_YYYY} The median aggregated CCSM3 projected annual fractional moisture in the entire soil column for year YYYY under scenario B1 by county. The source of the raw data was California Energy Commission (2008). The variables are smclB1_YYYY where YYYY is any of the years 2000, 2049 and 2099 (observed or projected), so smclB1_2069 is the projected median soil column moisture in that county under B1 in 2069.
\item{fireA2_YYYY} The median aggregated Centre National de Recherches Meteorologiques (CNRM) projected annual wildfire risk (observing 1 or more fires in the next 30 years) for each county in year YYYY under scenarios A2. The source of the raw data was California Energy Commission (2008). The variables are fireA2_YYYY and YYYY any of the years 2020, 2050 and 2085 (observed or projected), so fireA2_2085 is the projected median wildfire risk in that county under A2 in 2085.
#' \item{fireB1_YYYY} The median aggregated Centre National de Recherches Meteorologiques (CNRM) projected annual wildfire risk (observing 1 or more fires in the next 30 years) for each county in year YYYY under scenarios B1. The source of the raw data was California Energy Commission (2008). The variables are fireB1_YYYY and YYYY any of the years 2020, 2050 and 2085 (observed or projected), so fireB1_2085 is the projected median wildfire risk in that county under B1 in 2085.
}
}
\source{
Cooley, H., Moore, E., Heberger, M. and Allen, L. (2012) Social Vulnerability to Climate Change. California Energy Commission. Publication Number: CEC-500-2012-013 \url{https://pacinst.org/wp-content/uploads/sites/21/2014/04/social-vulnerability-climate-change-ca.pdf} and Heberger, M., Cooley, C., Herrera, P., Gleick, P. and Moore, E. (2009) The impacts of sea-level rise on the Californian coast. California Energy Commission. Publication Number: CEC-500-2009-024-F \url{https://pacinst.org/publication/the-impacts-of-sea-level-rise-on-the-california-coast/} (\url{https://www.pacinst.org/reports/sea_level_rise_data/Blk_fld.zip} for raw data) and California Energy Comission (2008) \url{https://cal-adapt.org/data/download/}
}
\description{
A dataset containing observed and projected indicators of climate change related natural hazards for 58 Californian counties. The values are actually the medians of the predicted distribution over spatial measurement points. It is a compiled data set from three sources and that has been aggregated to the county level. The projected data were derived under two different IPCC climate change scenarios (A2, the high emission scenario and B1, the moderate emission scenario). It further contains the county value of the California social vulnerability index.
}
\details{
Overall there are 50 indicators of natural hazard, one indicator of social vulnerability and 1 identifier of the county which were:
}
