% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.stops.R
\name{coef.stops}
\alias{coef.stops}
\title{S3 coef method for stops objects}
\usage{
\method{coef}{stops}(object, ...)
}
\arguments{
\item{object}{object of class stops}

\item{...}{addditional arguments}
}
\value{
a vector of hyperparmeters theta
}
\description{
S3 coef method for stops objects
}
