% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cops.R
\name{cop_cmdscale}
\alias{cop_cmdscale}
\title{COPS version of strain}
\usage{
cop_cmdscale(dis, theta = c(1, 1), weightmat = NULL, ndim = 2,
  init = NULL, ..., stressweight = 1, cordweight = 0.5, q = 1,
  minpts = 2, epsilon = 10, rang = NULL, verbose = 0, plot = FALSE,
  scale = TRUE, normed = TRUE)
}
\arguments{
\item{dis}{numeric matrix or dist object of a matrix of proximities}

\item{theta}{the theta vector of powers; this is either a scalar of the lambda transformation for the observed proximities, or a vector where the first is the kappa argument for the fitted distances (here internally fixed to 1) and the second the lambda argument. Defaults to 1 1}

\item{weightmat}{(optional) a matrix of nonnegative weights}

\item{ndim}{number of dimensions of the target space}

\item{init}{(optional) initial configuration}

\item{...}{additional arguments to be passed to the fitting procedure}

\item{stressweight}{weight to be used for the fit measure; defaults to 1}

\item{cordweight}{weight to be used for the cordillera; defaults to 0.5}

\item{q}{the norm of the corrdillera; defaults to 1}

\item{minpts}{the minimum points to make up a cluster in OPTICS; defaults to 2}

\item{epsilon}{the epsilon parameter of OPTICS, the neighbourhood that is checked; defaults to 10}

\item{rang}{range of the distances (min distance minus max distance). If NULL (default) the cordillera will be normed to each configuration's maximum distance, so an absolute value of goodness-of-clusteredness.}

\item{verbose}{numeric value hat prints information on the fitting process; >2 is extremely verbose}

\item{plot}{plot the cordillera}

\item{scale}{should the configuration be scaled to mean=0 and sd=1? Defaults to TRUE}

\item{normed}{should the cordillera be normed; defaults to TRUE}
}
\value{
A list with the components
\itemize{
        \item stress: the stress
        \item stress.m: default normalized stress
        \item coploss: the weighted loss value
        \item OC: the Optics cordillera value
        \item parameters: the parameters used for fitting (kappa, lambda)
        \item fit: the returned object of the fitting procedure
        \item cordillera: the cordillera object
}
}
\description{
COPS version of strain
}
\keyword{multivariate}

