% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cops.R
\name{coplossMin}
\alias{coplossMin}
\title{Fitting a COPS Model.}
\usage{
coplossMin(delta, kappa = 1, lambda = 1, nu = 1, theta = c(kappa,
  lambda, nu), weightmat = 1 - diag(nrow(delta)), ndim = 2, init = NULL,
  stressweight = 1, cordweight, q = 1, minpts = 2, epsilon = 10,
  rang = NULL, optimmethod = c("Nelder-Mead", "Newuoa"), verbose = 0,
  scale = TRUE, normed = TRUE, eps = 1e-07, itmax = 1e+05, ...)
}
\arguments{
\item{delta}{numeric matrix or dist object of a matrix of proximities}

\item{kappa}{power transformation for fitted distances}

\item{lambda}{power transformation for proximities}

\item{nu}{power transformation for weights}

\item{theta}{the theta vector of powers; the first is kappa (for the fitted distances if it exists), the second lambda (for the observed proximities if it exist), the third is nu (for the weights if it exists) . If less than three elements are is given as argument, it will be recycled. Defaults to 1 1 1. Will override any kappa, lmabda, nu parameters if they are given and do not match}

\item{weightmat}{(optional) a matrix of nonnegative weights; defaults to 1 for all off diagonals}

\item{ndim}{number of dimensions of the target space}

\item{init}{(optional) initial configuration}

\item{stressweight}{weight to be used for the fit measure; defaults to 1}

\item{cordweight}{weight to be used for the cordillera; if missing gets estimated from the initial configuration as}

\item{q}{the norm of the corrdillera; defaults to 1}

\item{minpts}{the minimum points to make up a cluster in OPTICS; defaults to 2}

\item{epsilon}{the epsilon parameter of OPTICS, the neighbourhood that is checked; defaults to 10}

\item{rang}{range of the minimum reachabilities to be considered. If missing it is found from the initial configuration by taking 1.5 times the maximal minimum reachability of the initial fit. If NULL it will be normed to each configuration's minimum and maximum distance, so an absolute value of goodness-of-clusteredness. Note that the latter is not necessarily desirable when comparing configurations for their relative clusteredness. See also \code{\link{cordillera}}}

\item{optimmethod}{What optimizer to use? Defaults to NEWUOA, Nelder-Mead is also supported.}

\item{verbose}{numeric value hat prints information on the fitting process; >2 is very verbose}

\item{scale}{should the configuration be scaled to mean=0 and sd=1 for calculating the cordillera? Defaults to TRUE}

\item{normed}{should the cordillera be normed; defaults to TRUE}

\item{eps}{accuracy, defulats to 1e-12}

\item{itmax}{maximum number of iterations. Defaults to 100000}

\item{...}{additional arguments to be passed to the optimization procedure}
}
\value{
A list with the components
        \itemize{
        \item coploss: the weighted loss value
        \item OC: the Optics cordillera
        \item optim: the object returned from the optimization procedure
        \item stress: the stress
        \item stress.m: default normalized stress
        \item parameters: the parameters used for fitting (kappa, lambda)
        \item fit: the returned object of the fitting procedure
        \item cordillera: the cordillera object
}
}
\description{
Minimizing Coploss for a clustered Power Stress MDS configuration with given hyperparameters theta.
}
\examples{
\donttest{ 
dis<-as.matrix(smacof::kinshipdelta)

#Coploss with equal weight to stress and cordillera 
res1<-coplossMin(dis,stressweight=0.5,cordweight=0.5) 
res1
summary(res1)
plot(res1)
}

}
\keyword{clustering}
\keyword{multivariate}

