% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerstress.R
\name{doubleCenter}
\alias{doubleCenter}
\title{Double centering of a matrix}
\usage{
doubleCenter(x)
}
\arguments{
\item{x}{numeric matrix}
}
\value{
the double centered matrix
}
\description{
Double centering of a matrix
}
