% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structurednessindices.R
\name{knn_dist}
\alias{knn_dist}
\title{calculate k nearest neighbours from a distance matrix}
\usage{
knn_dist(dis, k)
}
\arguments{
\item{dis}{distance matrix}

\item{k}{number of nearest neighbours (Note that with a tie, the function returns the alphanumerically first one!)}
}
\description{
calculate k nearest neighbours from a distance matrix
}
