% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{match_partial_ignorecase_nopunct}
\alias{match_partial_ignorecase_nopunct}
\title{function for lookup that partially matched and ignores cases and punctuation}
\usage{
match_partial_ignorecase_nopunct(lookup, entry, ...)
}
\arguments{
\item{lookup}{the lookup string}

\item{entry}{the registry entry}

\item{...}{additional arguments to pmatch}
}
\description{
function for lookup that partially matched and ignores cases and punctuation
}
