% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/powerstress.R, R/stops.R
\name{mkPower}
\alias{mkPower}
\title{MakePower}
\usage{
mkPower(x, theta)

mkPower(x, theta)
}
\arguments{
\item{x}{matrix}

\item{r}{numeric (power)}

\item{x}{matrix}

\item{r}{numeric (power)}
}
\description{
MakePower

MakePower
}

