% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{plot.cmdscaleE}
\alias{plot.cmdscaleE}
\title{S3 plot method for cmdscaleE}
\usage{
\method{plot}{cmdscaleE}(
  x,
  plot.type = c("confplot"),
  plot.dim = c(1, 2),
  col,
  label.conf = list(label = TRUE, pos = 3, col = 1, cex = 0.8),
  identify = FALSE,
  type = "p",
  pch = 20,
  asp = 1,
  main,
  xlab,
  ylab,
  xlim,
  ylim,
  legpos,
  ...
)
}
\arguments{
\item{x}{cmdscaleE object}

\item{plot.type}{type of plot}

\item{plot.dim}{dimensions used for plotting}

\item{col}{color}

\item{label.conf}{list of label options}

\item{identify}{boolean flag for interactively identify points}

\item{type}{type of plot}

\item{pch}{plotting character}

\item{asp}{aspect ratio (defaults to 1)}

\item{main}{main title}

\item{xlab}{label of x axis}

\item{ylab}{label of y axis}

\item{xlim}{limits of x axis}

\item{ylim}{limits of y axis}

\item{legpos}{position of legend}

\item{...}{additional arguments passed to plot}
}
\value{
No return value, just plots a 'cmdscaleE' object.
}
\description{
S3 plot method for cmdscaleE
}
\details{
This function duplicates the plot method for smacof so it can be used with cmdscaleE objects. See \code{\link[smacof]{plot.smacof}} for the arguments.
}
