% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cops.R
\name{plot.cops}
\alias{plot.cops}
\title{S3 plot method for cops objects}
\usage{
\method{plot}{cops}(x, plot.type = c("confplot", "reachplot", "Shepard",
  "resplot", "stressplot", "bubbleplot"), ...)
}
\arguments{
\item{x}{an object of class cops}

\item{plot.type}{String indicating which type of plot to be produced: "confplot", "reachplot", "resplot", "Shepard", "stressplot" (see details)}

\item{...}{Further plot arguments passed: see 'plot.smacof' and 'plot' for detailed information.

Details:

- Configuration plot (plot.type = "confplot"): Plots the MDS configurations.
- Reachability plot (plot.type = "confplot"): Plots the OPTICS reachability plot and the OPTICS cordillera
- Residual plot (plot.type = "resplot"): Plots the dissimilarities against the fitted distances.
- Shepard diagram (plot.type = "Shepard"): Diagram with the observed dissimilarities against the fitted distances including (isotonic) regression line.
- Stress decomposition plot (plot.type = "stressplot", only for SMACOF objects in $fit): Plots the stress contribution in of each observation. Note that it rescales the stress-per-point (SPP) from the corresponding smacof function to percentages (sum is 100). The higher the contribution, the worse the fit.
- Bubble plot (plot.type = "bubbleplot", only available for SMACOF objects $fit): Combines the configuration plot with the point stress contribution. The larger the bubbles, the better the fit.}
}
\description{
S3 plot method for cops objects
}

