% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cordillera.R
\name{plot.cordillera}
\alias{plot.cordillera}
\title{Plot method for OPTICS cordilleras}
\usage{
\method{plot}{cordillera}(x, colbp = "lightgrey", coll = "black",
  liwd = 1.5, ...)
}
\arguments{
\item{x}{an object of class cordillera}

\item{colbp}{color of the barplot.}

\item{coll}{color of the cordillera line}

\item{liwd}{width of the cordillera line}

\item{...}{additional arguments passed to barplot or lines}
}
\description{
Plots the reachability plot and adds the cordillera to it (as a line). In this plot the cordillera is proportional to the real value.
}

