% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/powerstress.R
\name{plot.smacofP}
\alias{plot.smacofP}
\title{S3 plot method for smacofP objects}
\usage{
\method{plot}{smacofP}(x, plot.type = "confplot", plot.dim = c(1, 2), main,
  xlab, ylab, xlim, ylim, col, ...)
}
\arguments{
\item{x}{an object of class cops}

\item{plot.type}{String indicating which type of plot to be produced: "confplot", "reachplot", "resplot", "Shepard", "stressplot","NLShepard" (see details)}

\item{plot.dim}{dimesnions to be ploted in confplot; defaults to c(1, 2)}

\item{main}{plot title}

\item{xlab}{label of x axis}

\item{ylab}{label of y axis}

\item{xlim}{scale of x axis}

\item{ylim}{scale of y axis}

\item{col}{vector of colors for the points}

\item{...}{Further plot arguments passed: see 'plot.smacof' and 'plot' for detailed information.

Details:

- Configuration plot (plot.type = "confplot"): Plots the MDS configurations.
- Reachability plot (plot.type = "confplot"): Plots the OPTICS reachability plot and the OPTICS cordillera
- Residual plot (plot.type = "resplot"): Plots the dissimilarities against the fitted distances.
- Linearized Shepard diagram (plot.type = "Shepard"): Diagram with the transformed observed dissimilarities against the transformed fitted distance as well as loess smooth and a least squares line.
- Nonlinear Shepard diagram (plot.type = "NLShepard"): Diagram with the observed dissimilarities (lighter) and the transformed observed dissimilarities (darker) against the fitted distances together with loess smoothing lines
- Stress decomposition plot (plot.type = "stressplot", only for SMACOF objects in $fit): Plots the stress contribution in of each observation. Note that it rescales the stress-per-point (SPP) from the corresponding smacof function to percentages (sum is 100). The higher the contribution, the worse the fit.
- Bubble plot (plot.type = "bubbleplot", only available for SMACOF objects $fit): Combines the configuration plot with the point stress contribution. The larger the bubbles, the better the fit.}
}
\description{
S3 plot method for smacofP objects
}

