% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.stops.R
\name{plot.stops}
\alias{plot.stops}
\title{S3 plot method for stops objects}
\usage{
\method{plot}{stops}(x, plot.type = "confplot", main, asp = 1, ...)
}
\arguments{
\item{x}{an object of class stops}

\item{plot.type}{String indicating which type of plot to be produced: "confplot", "resplot", "Shepard", "stressplot", "bubbleplot" (see details)}

\item{main}{the main title of the plot}

\item{asp}{aspect ratio of x/y axis; defaults to 1; setting to 1 will lead to an accurate represenation of the fitted distances.}

\item{...}{Further plot arguments passed: see 'plot.smacof' and 'plot' for detailed information.

Details:
See plot.smacofP}
}
\value{
no return value, just plots
}
\description{
S3 plot method for stops objects
}
