% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{plot3d.cmdscaleE}
\alias{plot3d.cmdscaleE}
\title{S3 plot3d method for class cmdscaleE}
\usage{
\method{plot3d}{cmdscaleE}(
  x,
  plot.dim = c(1, 2, 3),
  xlab,
  ylab,
  zlab,
  col,
  main,
  bgpng = NULL,
  ax.grid = TRUE,
  sphere.rgl = FALSE,
  ...
)
}
\arguments{
\item{x}{object of class cmdscaleE}

\item{plot.dim}{vector of length 3 with dimensions to be plotted}

\item{xlab}{label of x axis}

\item{ylab}{label of y axis}

\item{zlab}{label of z axis}

\item{col}{color of the text labels}

\item{main}{plot title}

\item{bgpng}{Background image from rgl library; 'NULL' for white background}

\item{ax.grid}{If 'TRUE', axes grid is plotted.}

\item{sphere.rgl}{If 'TRUE', rgl sphere (background) is plotted.}

\item{...}{Further plot arguments passed: see 'plot3d' in package 'rgl' for detailed information.}
}
\value{
No return value, just plots a 'cmdscale' object.
}
\description{
This methods produces a dynamic 3D configuration plot.
}
