% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{plot3dstatic.cmdscale}
\alias{plot3dstatic.cmdscale}
\title{3D plots: plot3dstatic method for class cmdscale}
\usage{
\method{plot3dstatic}{cmdscale}(x, plot.dim = c(1, 2, 3), main, xlab,
  ylab, zlab, col, ...)
}
\arguments{
\item{x}{object of class cmdscale}

\item{plot.dim}{vector of length 3 with dimensions to be plotted}

\item{main}{plot title}

\item{xlab}{label of x axis}

\item{ylab}{label of y axis}

\item{zlab}{label of z axis}

\item{col}{color of the text labels}

\item{...}{Further plot arguments passed: see 'scatterplot3d' in package 'scatterplot3d' for detailed information.}
}
\description{
This methods produces a static 3D configuration plot.
}
