% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{print.cmdscale}
\alias{print.cmdscale}
\title{S3 print method for cmdscale}
\usage{
\method{print}{cmdscale}(x, ...)
}
\arguments{
\item{x}{cmdscale object}

\item{...}{additional arguments}
}
\value{
No return value, just prints.
}
\description{
S3 print method for cmdscale
}
