% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cordillera.R
\name{print.cordillera}
\alias{print.cordillera}
\title{Print method for OPTICs cordilleras}
\usage{
\method{print}{cordillera}(x, ...)
}
\arguments{
\item{x}{an object of class optics}

\item{...}{additional arguments passed to print}
}
\description{
Prints the raw and normalized cordillera
}

