% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optics.R
\name{print.optics}
\alias{print.optics}
\title{Print method for OPTICS results}
\usage{
\method{print}{optics}(x, ...)
}
\arguments{
\item{x}{an object of class optics}

\item{...}{additional arguments passed to print}
}
\value{
a data frame with the observation order and the reachabilities (invisible)
}
\description{
Prints the object ordering and the reachabilities as directly read in from ELKI.
}

