% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optics.R
\name{print.opticse}
\alias{print.opticse}
\title{Print method for OPTICS results}
\usage{
\method{print}{opticse}(x, ...)
}
\arguments{
\item{x}{an object of class optics}

\item{...}{additional arguments passed to print}
}
\value{
a data frame with the observation order and the reachabilities (invisible)
}
\description{
Prints the object ordering and the reachabilities as directly read in from ELKI.
}

