% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{print.sammon}
\alias{print.sammon}
\title{S3 print method for sammon objects}
\usage{
\method{print}{sammon}(x, ...)
}
\arguments{
\item{x}{cmdscale object}

\item{...}{additional arguments}
}
\value{
No return value, just prints.
}
\description{
S3 print method for sammon objects
}
