% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.stops.R
\name{print.stops}
\alias{print.stops}
\title{S3 print method for stops objects}
\usage{
\method{print}{stops}(x, ...)
}
\arguments{
\item{x}{stops object}

\item{...}{additional arguments}
}
\value{
no return value, just prints
}
\description{
S3 print method for stops objects
}
