% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optics.R
\name{print.summary.opticse}
\alias{print.summary.opticse}
\title{Print method for OPTICS summary
Displays summaries of the reachability plot. Currently its the five points summary of the reachabilities and a stem and leaf display. The latter should not be confused with the reachability plot. If you need the latter, use plot()}
\usage{
\method{print}{summary.opticse}(x, fiven = TRUE, stemd = TRUE, ...)
}
\arguments{
\item{x}{an object of class summary.optics}

\item{fiven}{should the 5 point summary be printed. Default is TRUE.}

\item{stemd}{should the stema dn leaf plot be printed. Default is TRUE.}

\item{...}{additional arguments passed to \code{\link{stem}}}
}
\description{
Print method for OPTICS summary
Displays summaries of the reachability plot. Currently its the five points summary of the reachabilities and a stem and leaf display. The latter should not be confused with the reachability plot. If you need the latter, use plot()
}

