% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerstress.R
\name{print.summary.smacofP}
\alias{print.summary.smacofP}
\title{S3 print method for summary.smacofP}
\usage{
\method{print}{summary.smacofP}(x, ...)
}
\arguments{
\item{x}{object of class summary.smacofP}

\item{...}{additional arguments}
}
\value{
No return value, just prints a 'summary.smacofP'
}
\description{
S3 print method for summary.smacofP
}
