% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops.R
\name{print.summary.stops}
\alias{print.summary.stops}
\title{S3 print method for summary.stops}
\usage{
\method{print}{summary.stops}(x, ...)
}
\arguments{
\item{x}{object of class summary.stops}

\item{...}{additional arguments}
}
\value{
no return value, just prints
}
\description{
S3 print method for summary.stops
}
