% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops.R
\name{residuals.stops}
\alias{residuals.stops}
\title{S3 residuals method for stops}
\usage{
\method{residuals}{stops}(object, ...)
}
\arguments{
\item{object}{object of class stops}

\item{...}{addditional arguments}
}
\value{
a vector of residuals (observed minus fitted distances)
}
\description{
S3 residuals method for stops
}
