% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cops.R
\name{shrinkB}
\alias{shrinkB}
\title{Finding the shrinkage matrix for COPS-0}
\usage{
shrinkB(x, q = 1, minpts = 2, epsilon = 10, rang = NULL,
  scaleB = TRUE, normed = 2, ...)
}
\arguments{
\item{x}{numeric matrix}

\item{q}{the norm to be used}

\item{minpts}{the minimum points to make up a cluster in OPTICS; defaults to ndim+1}

\item{epsilon}{the epsilon parameter of OPTICS, the neighbourhood that is checked; defaults to 10}

\item{rang}{range of the minimum reachabilities to be considered. If missing it is found from the initial configuration.}

\item{scaleB}{should the X matrix be scaled before calculating the shrinkage weights; defaults to TRUE (which is sensible for relative shrinkage)}

\item{normed}{should the reachability differences be normed? (1=to sup Gamma 2 = to dmax)?}

\item{...}{additional arguments to be passed to the OPTICS algorithm procedure}
}
\description{
Finding the shrinkage matrix for COPS-0
}
