% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stops.R
\name{stop_flexsmacof}
\alias{stop_flexsmacof}
\title{STOPS versions of flexsmacof models (models with a parametric f() transformation to be determined)}
\usage{
stop_flexsmacof(dis, transformation = mkPower, theta = c(1, 1), ndim = 2,
  weightmat = NULL, init = NULL, ..., structures = c("clusteredness",
  "linearity"), stressweight = 1, strucweight = rep(1/length(structures),
  length(structures)), strucpars)
}
\arguments{
\item{dis}{numeric matrix or dist object of a matrix of proximities}

\item{transformation}{function to transform the proximities or distances; need to be parameterized by theta}

\item{theta}{the theta vector of transformations}

\item{ndim}{number of dimensions of the target space}

\item{weightmat}{(optional) a matrix of nonnegative weights}

\item{init}{(optional) initial configuration}

\item{...}{additional arguments to be passed to the fitting}

\item{structures}{which structuredness indices to be included in the loss}

\item{stressweight}{weight to be used for the fit measure; defaults to 1}

\item{strucweight}{weight to be used for the structuredness indices; ; defaults to 1/#number of structures}

\item{q}{the norm of the corrdillera; defaults to 1}

\item{minpts}{the minimum points to make up a cluster in OPTICS; defaults to 2}

\item{epsilon}{the epsilon parameter of OPTICS, the neighbourhood that is checked; defaults to 10}

\item{rang}{range of the distances (min distance minus max distance)}

\item{verbose}{numeric value hat prints information on the fitting process; >2 is extremely verbose}

\item{plot}{plot the cordillera}

\item{normed}{should the cordillera be normed; defaults to TRUE}

\item{scale}{should the configuration be scaled to mean=0 and sd=1? Defaults to TRUE}
}
\value{
A list with the components
   \itemize{
        \item{stress:} the stress
        \item{stress.m:} default normalized stress
        \item{stoploss:} the weighted loss value
        \item{indices:} the values of the structuredness indices
        \item{parameters:} the parameters used for fitting
        \item{fit:} the returned object of the fitting procedure
        \item{indobj:} the index objects
}
}
\description{
STOPS versions of flexsmacof models (models with a parametric f() transformation to be determined)
}
\keyword{multivariate}

