% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops.R
\name{stop_isomap}
\alias{stop_isomap}
\title{STOPS version of isomap.}
\usage{
stop_isomap(dis, theta = 3, weightmat = NULL, ndim = 2, init = NULL,
  stressweight = 1, structures = c("cclusteredness", "clinearity",
  "cdependence", "cmanifoldness", "cassociation", "cnonmonotonicity",
  "cfunctionality", "ccomplexity", "cfaithfulness", "cregularity"),
  strucweight = rep(1/length(structures), length(structures)), strucpars,
  verbose = 0, type = c("additive", "multiplicative"))
}
\arguments{
\item{dis}{numeric matrix or dist object of a matrix of proximities}

\item{theta}{the number of shortest dissimilarities retained for a point (nearest neighbours), the isomap parameter. Defaults to 3.}

\item{weightmat}{(optional) a matrix of nonnegative weights}

\item{ndim}{number of dimensions of the target space}

\item{init}{(optional) initial configuration}

\item{stressweight}{weight to be used for the fit measure; defaults to 1}

\item{structures}{which structuredness indices to be included in the loss}

\item{strucweight}{weight to be used for the structuredness indices; ; defaults to 1/#number of structures}

\item{strucpars}{the parameters for the structuredness indices}

\item{verbose}{numeric value hat prints information on the fitting process; >2 is extremely verbose}

\item{type}{How to construct the target function for the multi objective optimization? Either 'additive' (default) or 'multiplicative'}
}
\value{
A list with the components
   \itemize{
        \item{stress:} Not really stress but 1-GOF where GOF is the first element returned from cmdscale (the sum of the first ndim absolute eigenvalues divided by the sum of all absolute eigenvalues).
        \item{stress.m:} default normalized stress (sqrt explicitly normalized stress; really the stress this time)
        \item{stoploss:} the weighted loss value
        \item{indices:} the values of the structuredness indices
        \item{parameters:} the parameters used for fitting 
        \item{fit:} the returned object of the fitting procedure
        \item{indobj:} the index objects
}
}
\description{
Currently this version is a bit less flexible than the vegan one, as the only allowed parameter for isomap is the theta (k in isomap, no epsilon) and the shortest path is always estimated with argument "shortest". Also note that fragmentedOK is always set to TRUE which means that for theta that is too small only the largest conected group will be analyzed. If that's not wanted just set the theta higher.
}
\keyword{multivariate}
