% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stops.R
\name{stop_sammon2}
\alias{stop_sammon2}
\title{STOPS versions of Sammon mapping models (via smacofSym)}
\usage{
stop_sammon2(dis, theta = c(1, 1, -1), ndim = 2, weightmat = NULL,
  init = NULL, ..., stressweight = 1, structures = c("cclusteredness",
  "clinearity", "cdependence", "cmanifoldness", "cassociation",
  "cnonmonotonicity", "cfunctionality", "ccomplexity", "cfaithfulness"),
  strucweight = rep(1/length(structures), length(structures)), strucpars,
  verbose = 0, type = c("additive", "multiplicative"),
  stresstype = "default")
}
\arguments{
\item{dis}{numeric matrix or dist object of a matrix of proximities}

\item{theta}{the theta vector of powers; this is either a scalar of the lambda transformation for the observed proximities, or a vector where the first is the kappa argument for the fitted distances (here internally fixed to 1) and the second the lambda argument, the thrid the nu argiument (fixed to -1). Defaults to 1 1 -1.}

\item{ndim}{number of dimensions of the target space}

\item{weightmat}{(optional) a matrix of nonnegative weights}

\item{init}{(optional) initial configuration}

\item{...}{additional arguments to be passed to the fitting procedure}

\item{stressweight}{weight to be used for the fit measure; defaults to 1}

\item{structures}{which structuredness indices to be included in the loss}

\item{strucweight}{weight to be used for the structuredness indices; ; defaults to 1/#number of structures}

\item{strucpars}{the parameters for the structuredness indices}

\item{verbose}{numeric value hat prints information on the fitting process; >2 is extremely verbose}

\item{type}{How to construct the target function for the multi objective optimization? Either 'additive' (default) or 'multiplicative'}

\item{stresstype}{which stress to report. Only takes smacofs default stress currrently.}
}
\value{
A list with the components
   \itemize{
        \item{stress:} the stress
        \item{stress.m:} default normalized stress
        \item{stoploss:} the weighted loss value
        \item{indices:} the values of the structuredness indices
        \item{parameters:} the parameters used for fitting
        \item{fit:} the returned object of the fitting procedure
        \item{indobj:} the index objects
}
}
\description{
STOPS versions of Sammon mapping models (via smacofSym)
}
\keyword{multivariate}

