% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{summary.cmdscale}
\alias{summary.cmdscale}
\title{S3 summary method for cmdscale}
\usage{
\method{summary}{cmdscale}(object, ...)
}
\arguments{
\item{object}{object of class cmdscale}

\item{...}{additional arguments}
}
\value{
No return value, just prints.
}
\description{
S3 summary method for cmdscale
}
