% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cordillera.R
\name{summary.cordillera}
\alias{summary.cordillera}
\title{Summary method for OPTICS cordilleras}
\usage{
\method{summary}{cordillera}(object, ...)
}
\arguments{
\item{object}{an object of class optics}

\item{...}{additional arguments passed to summary or stem}
}
\description{
Displays a summary of the cordillera (OC), including information on OPTICS, the Raw OC, the normalisation factor and  the normalized cordillera
}

