% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optics.R
\name{summary.optics}
\alias{summary.optics}
\title{Summary method for OPTICs results}
\usage{
\method{summary}{optics}(object, fiven = TRUE, stemd = TRUE, ...)
}
\arguments{
\item{object}{an object of class optics}

\item{fiven}{flags whether the 5-point summary should be printed. Defaults to TRUE.}

\item{stemd}{flags whether the stem an leaf display should be printed. Defaults to TRUE.}

\item{...}{additional arguments passed to summary or stem}
}
\value{
invisibly returns the reachabilities as a numeric vector
}
\description{
Displays summaries of the reachability plot. Currently its the five points summary of the reachabilities and a stem and leaf display. The latter should not be confused with the reachability plot. If you need the altter, use plot().
}

