% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{summary.sammon}
\alias{summary.sammon}
\title{S3 summary method for sammon}
\usage{
\method{summary}{sammon}(object, ...)
}
\arguments{
\item{object}{object of class sammon}

\item{...}{additional arguments}
}
\value{
No return value, just prints.
}
\description{
S3 summary method for sammon
}
