% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerstress.R
\name{summary.smacofP}
\alias{summary.smacofP}
\title{S3 summary method for smacofP}
\usage{
\method{summary}{smacofP}(object, ...)
}
\arguments{
\item{object}{object of class smacofP}

\item{...}{additional arguments}
}
\value{
an object of class summary.smacofP
}
\description{
S3 summary method for smacofP
}
