% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops.R
\name{summary.stops}
\alias{summary.stops}
\title{S3 summary method for stops}
\usage{
\method{summary}{stops}(object, ...)
}
\arguments{
\item{object}{object of class stops}

\item{...}{addditional arguments}
}
\value{
object of class 'summary.stops'
}
\description{
S3 summary method for stops
}
