% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Swissroll-data.R
\name{Swissroll}
\alias{Swissroll}
\title{Swiss roll}
\format{
A data frame with 150 rows and 4 columns
}
\usage{
data(Swissroll)
}
\description{
A swiss roll data example where 150 data points are arranged on a swiss roll embedded in a 3D space.
}
\details{
A data frame with the variables (columns)
\itemize{
\item x The x axis coordinate for each point
\item y The y axis coordinate for each point
\item z The z axis coordinate for each point
\item col a color code for each point with points along the y axis having the same color (based on the viridis palette) 
}
}
