% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerstress.R
\name{torgerson}
\alias{torgerson}
\title{Torgerson scaling}
\usage{
torgerson(delta, p = 2)
}
\arguments{
\item{delta}{symmetric, numeric matrix of distances}

\item{p}{target space dimensions}
}
\value{
a n x p matrix (the configuration)
}
\description{
Torgerson scaling
}
\examples{
dis<-as.matrix(smacof::kinshipdelta)
res<-torgerson(dis)
}
