% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcstress.R
\name{bcStressMin}
\alias{bcStressMin}
\title{An MDS version for minimizing BoxCox Stress (Chen & Buja 2013)}
\usage{
bcStressMin(delta, init = NULL, verbose = 0, ndim = 2, lambda = 1,
  mu = 1, nu = 0, itmax = 10000)
}
\arguments{
\item{delta}{dissimilarity or distance matrix}

\item{init}{initial configuration. If NULL a classical scaling solution is used.}

\item{verbose}{prints progress if > 3.}

\item{ndim}{the dimension of the configuration}

\item{lambda}{lambda parameter}

\item{mu}{mu parameter}

\item{nu}{the nu parameter}

\item{itmax}{number of optimizing iterations, defaults to 10000.}
}
\description{
Based on code by Lisha Chen.
}
\examples{
dis<-smacof::kinshipdelta
res<-bcStressMin(as.matrix(dis),mu=2,lambda=1.5,nu=0)
res
summary(res)
plot(res)

}
