% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcstress.R
\name{bcStressMin}
\alias{bcStressMin}
\title{An MDS version for minimizing BoxCox Stress (Chen & Buja 2013)}
\usage{
bcStressMin(delta, init = NULL, verbose = 0, ndim = 2, mu = 1,
  lambda = 1, nu = 0, itmax = 10000, addD0 = 1e-04)
}
\arguments{
\item{delta}{dissimilarity or distance matrix}

\item{init}{initial configuration. If NULL a classical scaling solution is used.}

\item{verbose}{prints progress if > 3.}

\item{ndim}{the dimension of the configuration}

\item{mu}{mu parameter. Should be 0 or larger for everything working ok. If mu<0 it works but the model is strange and normalized stress tends towards 0 regardless of fit. Use normalized stress at your own risk in that case.}

\item{lambda}{lambda parameter. Must be larger than 0.}

\item{nu}{the nu parameter.}

\item{itmax}{number of optimizing iterations, defaults to 10000.}

\item{addD0}{a small number that's added for D(X)=0 for numerical evaluation of worst fit (numerical reasons, see details). If addD0=0 the normalized stress for mu!=0 and mu+lambda!=0 is correct, but will give useless normalized stress for mu=0 or mu+lambda!=0.}
}
\description{
An MDS version for minimizing BoxCox Stress (Chen & Buja 2013)
}
\details{
For numerical reasons with certain parameter combinations, the normalized stress uses a configuration as worst result where every d(X) is 0+addD0. The same number is not added to the delta so there is a small inaccuracy of the normalized stress (but negligible of min(delta)>>addD0). Also, for mu<0 or mu+lambda<0 the normalization can not generally be trusted (in the worst case of D(X)=0 one would have an 0^(-a)).
}
\examples{
dis<-smacof::kinshipdelta
res<-bcStressMin(as.matrix(dis),mu=2,lambda=1.5,nu=0)
res
summary(res)
plot(res)

}
\author{
Lisha Chen & Thomas Rusch
}
