% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/structurednessindices.R
\name{c_dependence}
\alias{c_dependence}
\title{c-dependence
calculates c-dependence as the distance correlation}
\usage{
c_dependence(confs, index = 1)
}
\arguments{
\item{confs}{a numeric matrix or data frame with two columns}

\item{index}{exponent on Euclidean distance, in (0,2]}
}
\description{
c-dependence
calculates c-dependence as the distance correlation
}
\examples{
x<-1:10
y<-2+3*x+rnorm(10)
confs<-cbind(x,y)
c_dependence(confs,1.5)
}

