% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structurednessindices.R
\name{c_inequality}
\alias{c_inequality}
\title{c-inequality
calculates c-inequality (as in an economic measure of inequality) as Pearsons coefficient of variation of the distance matrix
this is one of few c-structuredness indices not between 0 and 1, but 0 and infinity}
\usage{
c_inequality(confs)
}
\arguments{
\item{confs}{a numeric matrix or data frame}
}
\description{
c-inequality
calculates c-inequality (as in an economic measure of inequality) as Pearsons coefficient of variation of the distance matrix
this is one of few c-structuredness indices not between 0 and 1, but 0 and infinity
}
\examples{
x<-1:10
y<-2+3*x+rnorm(10)
z<- sin(y-x)
confs<-cbind(z,y,x)
c_inequality(confs)
}
