% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/structurednessindices.R
\name{c_linearity}
\alias{c_linearity}
\title{c-linearity
calculates c-linearity as the multiple correlation}
\usage{
c_linearity(confs, ...)
}
\arguments{
\item{confs}{a numeric matrix or data frame}

\item{...}{additional arguments to be passed to lm.fit}
}
\description{
c-linearity
calculates c-linearity as the multiple correlation
}
\examples{
x<-1:10
y<-2+3*x+rnorm(10)
confs<-cbind(x,y)
clinearity(confs)
}

