% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/structurednessindices.R
\name{c_linearity}
\alias{c_linearity}
\title{c-linearity
calculates c-linearity as the maximum multiple correlation}
\usage{
c_linearity(confs)
}
\arguments{
\item{confs}{a numeric matrix or data frame}
}
\description{
c-linearity
calculates c-linearity as the maximum multiple correlation
}
\examples{
x<-1:10
y<-2+3*x+rnorm(10)
z<- sin(y-x)
confs<-cbind(z,y,x)
c_linearity(confs)
}

