% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structurednessindices.R
\name{c_manifoldness}
\alias{c_manifoldness}
\title{c-manifoldness
calculates c-manifoldness as the highest maximal correlation coefficient (i.e., Pearson correlation of the ACE transformedvariables) of all pairwise combinations of two different columns in confs. If there is an NA (happens usually when the optimal transformation of any variable is a constant and therefore the covariance is 0 but also one of the sds in the denominator), it gets skipped.}
\usage{
c_manifoldness(confs)
}
\arguments{
\item{confs}{a numeric matrix or data frame with two columns}
}
\description{
c-manifoldness
calculates c-manifoldness as the highest maximal correlation coefficient (i.e., Pearson correlation of the ACE transformedvariables) of all pairwise combinations of two different columns in confs. If there is an NA (happens usually when the optimal transformation of any variable is a constant and therefore the covariance is 0 but also one of the sds in the denominator), it gets skipped.
}
\examples{
x<--100:100
y<-sqrt(100^2-x^2)
confs<-cbind(x,y)
c_manifoldness(confs)
}
