% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structurednessindices.R
\name{c_outlying}
\alias{c_outlying}
\title{c-outlying}
\usage{
c_outlying(conf)
}
\arguments{
\item{conf}{A numeric matrix.}
}
\description{
Measures the c-outlying structure
}
\examples{
delts<-smacof::kinshipdelta
conf3<-smacof::smacofSym(delts,ndim=3)$conf
c_outlying(conf3)
}
