% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cops.R
\name{cops}
\alias{cops}
\title{High Level COPS Function}
\usage{
cops(dis, variant = c("1", "2", "variant1", "variant2", "v1", "v2",
  "configuration", "profile", "coploss", "p-coploss"), ...)
}
\arguments{
\item{dis}{a dissimilarity matrix or a dist object}

\item{variant}{a character string specifying which variant of COPS to fit. Allowed is any of the following "1","2","Variant1","Variant2","v1","v2","configuration","profile","coploss","p-coploss". Defaults to "1".}

\item{...}{arguments to be passed to coplossMin (for Variant 1) or pcops (for Variant 2). See also \code{\link{coplossMin}} or \code{\link{pcops}}}
}
\value{
For Variant 1 see \code{\link{coplossMin}}, for Variant 2 see \code{\link{pcops}}
}
\description{
Minimizing coploss for a clustered MDS configuration. Allows to choose COPS Variant 1 (finding a configuration from coploss) and Variant 2 (finding hyperparameters for MDS models with power transformations). It is wrapper for coplossMin (Variant 1) and pcops (Variant 2).
}
\examples{
\donttest{
dis<-as.matrix(smacof::kinshipdelta)

#Coploss with equal weight to stress and cordillera 
res1<-cops(dis,stressweight=0.5,cordweight=0.5) 
res1
summary(res1)
plot(res1)

#classic mds (i.e. stressweight=1 and cordweight=0)
res2<-cops(dis,stressweight=1,cordweight=0) 
res2
summary(res2)
plot(res2)

#cordillera value of res1 and res 2 very close but res 2 is a bit more clustered
#the reason is the distance between sister and son

#procrustes adjusted
resadj<-conf_adjust(res2$fit$points,res1$fit$points)
plot(resadj$ref.conf) #res 2
plot(resadj$other.conf) #res 1

par(mfrow=c(1,2))
plot(res1,"reachplot")
plot(res2,"reachplot") 

#s-stress type coploss (i.e. kappa=2, lambda=2)
res3<-cops(dis,kappa=2,lambda=2,stressweight=0.5,cordweight=0.5) 
res3
summary(res3)
plot(res3)

#Sammon stress type coploss
ws<-weightmat=1/dis
diag(ws)<-1 
res4<-cops(dis,nu=-1,weightmat=ws,stressweight=0.5,cordweight=0.5) 
res4
summary(res4)
plot(res4)

#power-stress type coploss
ws<-weightmat=1/dis
diag(ws)<-1 
res5<-cops(dis,kappa=1.4,lambda=3,nu=-1,weightmat=ws,stressweight=0.5,cordweight=0.5) 
res5
summary(res5)
plot(res5)
}

}
\keyword{clustering}
\keyword{multivariate}

