% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cops.R
\name{cops}
\alias{cops}
\title{High Level COPS Function}
\usage{
cops(dis, variant = c("0", "1", "2", "Variant0", "Variant1", "Variant2", "v0",
  "v1", "v2", "COPS-0", "COPS-C", "P-COPS", "configuration-0",
  "configuration-c", "profile", "coploss-0", "coploss-c", "p-coploss", "COPS-P",
  "coploss-p"), ...)
}
\arguments{
\item{dis}{a dissimilarity matrix or a dist object}

\item{variant}{a character string specifying which variant of COPS to fit. Allowed is any of the following "0","1","2","Variant0","Variant1","Variant2","v0","v1","v2","COPS-0","COPS-C","P-COPS","configuration-0","configuration-c","profile","coploss-0","coploss-c","p-coploss". Defaults to "COPS-C".}

\item{...}{arguments to be passed to shrinkCoploss0 (for Variant 0) coplossMin (for Variant 1) or pcops (for Variant 2). See also \code{\link{shrinkCoploss0}} \code{\link{coplossMin}} or \code{\link{pcops}}}
}
\value{
For Variant 0 see \code{\link{shrinkCoploss0}}, Variant 1 see \code{\link{coplossMin}}, for Variant 2 see \code{\link{pcops}}
}
\description{
Minimizing coploss for a clustered MDS configuration. Allows to choose COPS-0 (finding a configuration from coploss with residual shrinkage to zero) and COPS-C (finding a configuration from coploss with cordillera penalty) and profile COPS (finding hyperparameters for MDS models with power transformations). It is wrapper for shrinkCoploss0, coplossMin and pcops.
}
\examples{
\donttest{
dis<-as.matrix(smacof::kinshipdelta)

#COPS-C with equal weight to stress and cordillera 
res1<-cops(dis,variant="COPS-C",stressweight=0.5,cordweight=0.5,minpts=2) 
res1
summary(res1)
plot(res1)

#classic mds (i.e. COPS-C with stressweight=1 and cordweight=0 or COPS-0 with cordweight=0)
res2<-cops(dis,variant="COPS-C",stressweight=1,cordweight=0,minpts=2) 
res2
summary(res2)
plot(res2)

#procrustes adjusted
resadj<-conf_adjust(res2$fit$conf,res1$fit$conf)
plot(resadj$ref.conf) #res 2
plot(resadj$other.conf) #res 1

par(mfrow=c(1,2))
plot(res1,"reachplot")
plot(res2,"reachplot") 


#COPS-0 to improve over an MDS result
res0<-powerStressFast(dis)
res2a<-cops(dis,variant="COPS-0",cordweight=1,q=2,init=res0$conf,minpts=2) 
res2a
summary(res2a)
plot(res2a)

resadj<-conf_adjust(res0$fit$conf,res2a$fit$conf)
plot(resadj$ref.conf) #res 0
plot(resadj$other.conf) #res 2a


#s-stress type coploss (i.e. kappa=2, lambda=2)
res3<-cops(dis,variant="COPS-C",kappa=2,lambda=2,stressweight=0.5,cordweight=0.5) 
res3
summary(res3)
plot(res3)

#Sammon stress type coploss
ws<-weightmat=1/dis
diag(ws)<-1 
res4<-cops(dis,variant="COPS-0",nu=-1,weightmat=ws,cordweight=0.5) 
res4
summary(res4)
plot(res4)

#power-stress type profile coploss
# search for optimal kappa and lambda between kappa=0.5,lambda=0.5 and kappa=2,lambda=5
# nu is fixed on -1
ws<-1/dis
diag(ws)<-1 
res5<-cops(dis,variant="P-COPS",loss="powerstress",theta=c(1.4,3,-1),lower=c(1,0.5,-1),upper=c(3,5,-1),weightmat=ws,stressweight=0.9,cordweight=0.1) 
res5
summary(res5)
plot(res5)
}

}
\keyword{clustering}
\keyword{multivariate}
