% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cops.R
\name{cops}
\alias{cops}
\title{High Level COPS Function}
\usage{
cops(...)
}
\arguments{
\item{...}{arguments to be passed to coplossMin. See also \code{\link{coplossMin}}}
}
\value{
A list with the components
        \itemize{
        \item coploss: the weighted loss value
        \item OC: the Optics cordillera
        \item optim: the object returned from the optimization procedure
        \item stress: the stress
        \item stress.m: default normalized stress
        \item parameters: the parameters used for fitting (kappa, lambda)
        \item fit: the returned object of the fitting procedure
        \item cordillera: the cordillera object
}
}
\description{
Minimizing Coploss for a clustered Power Stress MDS configuration with given hyperparameters theta. Calls coplossMin.
}
\examples{
\donttest{
dis<-as.matrix(smacof::kinshipdelta)

#Coploss with equal weight to stress and cordillera 
res1<-cops(dis,stressweight=0.5,cordweight=0.5) 
res1
summary(res1)
plot(res1)

#classic mds (i.e. stressweight=1 and cordweight=0)
res2<-cops(dis,stressweight=1,cordweight=0) 
res2
summary(res2)
plot(res2)

#cordillera value of res1 and res 2 very close but res 2 is a bit more clustered
#the reason is the distance between sister and son

#procrustes adjusted
resadj<-conf_adjust(res2$fit$points,res1$fit$points)
plot(resadj$ref.conf) #res 2
plot(resadj$other.conf) #res 1

par(mfrow=c(1,2))
plot(res1,"reachplot")
plot(res2,"reachplot") 

#s-stress type coploss (i.e. kappa=2, lambda=2)
res3<-cops(dis,kappa=2,lambda=2,stressweight=0.5,cordweight=0.5) 
res3
summary(res3)
plot(res3)

#Sammon stress type coploss
ws<-weightmat=1/dis
diag(ws)<-1 
res4<-cops(dis,nu=-1,weightmat=ws,stressweight=0.5,cordweight=0.5) 
res4
summary(res4)
plot(res4)

#power-stress type coploss
ws<-weightmat=1/dis
diag(ws)<-1 
res5<-cops(dis,kappa=1.4,lambda=3,nu=-1,weightmat=ws,stressweight=0.5,cordweight=0.5) 
res5
summary(res5)
plot(res5)
}

}
\keyword{clustering}
\keyword{multivariate}

