% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmds.R
\name{lmds}
\alias{lmds}
\title{An MDS version for local MDS (Chen & Buja 2006)}
\usage{
lmds(delta, init = NULL, ndim = 3, k = 2, tau = 1, itmax = 1000,
  verbose = 0)
}
\arguments{
\item{delta}{dissimilarity or distance matrix}

\item{init}{initial configuration. If NULL a classical scaling solution is used.}

\item{ndim}{the dimension of the configuration}

\item{k}{the k neighbourhood parameter parameter}

\item{tau}{the penalty parameter}

\item{itmax}{number of optimizing iterations, defaults to 10000.}

\item{verbose}{prints progress if > 3.}
}
\description{
An MDS version for local MDS (Chen & Buja 2006)
}
\examples{
dis<-smacof::kinshipdelta
res<- lmds(as.matrix(dis),k=2,tau=0.1)
res
summary(res)
plot(res)

}
\author{
Lisha Chen & Thomas Rusch
}
