% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/powerstress.R
\name{powerStressMin}
\alias{powerStressMin}
\title{Power stress}
\usage{
powerStressMin(delta, kappa = 1, lambda = 1, nu = 1, lambdamax = lambda,
  weightmat = 1 - diag(nrow(delta)), init = NULL, ndim = 2, eps = 1e-10,
  itmax = 1e+05, verbose = FALSE, stresstype = stressen1)
}
\arguments{
\item{delta}{dist object or a symmetric, numeric data.frame or matrix of distances}

\item{kappa}{power of the transformation of the fitted distances; defaults to 1}

\item{lambda}{the power of the transformation of the proximities; defaults to 1}

\item{nu}{the power of the transformation for weightmat; defaults to 1}

\item{lambdamax}{the maximum power of the transformation of the proximities if there are more than one lambda - an upper bound idea; defaults to lambda}

\item{weightmat}{a matrix of finite weights}

\item{init}{starting configuration}

\item{ndim}{dimension of the configuration; defaults to 2}

\item{eps}{numeric accuracy of the iteration}

\item{itmax}{maximum number of iterations}

\item{verbose}{should iteration output be printed; if > 1 then yes}

\item{stresstype}{the stress type to be reported by default; defaults to stress-1 for smacof compatibility}
}
\value{
a smacofP object (inheriting form smacofB, see \code{\link{smacofSym}}). It is a list with the components
\itemize{
\item delta: Observed dissimilarities, not normalized
\item obsdiss: Observed dissimilarities, normalized to sum*w*delta=1
\item confdiss: Configuration dissimilarities
\item conf: Matrix of fitted configuration
\item stress: Default stress
\item spp: Stress per point (based on stress.en)
\item ndim: Number of dimensions
\item model: Name of smacof model
\item niter: Number of iterations
\item nobj: Number of objects
\item type: Type of MDS model
}
and some additional components
\itemize{
\item gamma: The majorizing function at convergence
\item stress.m: default stress for the COPS and STOP defaults to square root of the explicitly normalized stress on the normalized, transformed dissimilarities
\item stress.r: raw stress on the observed, transformed dissimilarities
\item stress.s: square root of the explicitly normalized stress on the observed, transformed dissimilarities
\item stress.n: explicitly normalized stress on the observed, transformed dissimilarities
\item stress.1: implicitly normalized stress on the observed, transformed dissimilarities
\item stress.e: raw stress on the normalized, transformed dissimilarities
\item stress.en: stress on the normalized, transformed dissimilarities and normalized transformed distances
\item stress.en1: stress 1 on the normalized, transformed dissimilarities and normalized transformed distances; this is reported by the print ,ethod
\item stress.e1: implicitly normalized stress on the normalized, transformed dissimilarities
\item deltaorig: observed, untransformed dissimilarities
\item weightmat: weighting matrix
}
}
\description{
An implemenation to minimize power stress
}
\examples{
dis<-smacof::kinshipdelta
res<-powerStressMin(as.matrix(dis),kappa=2,lambda=1.5)
res
summary(res)
plot(res)
}
\seealso{
\code{\link{smacofSym}}
}

