% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops.R
\name{stop_powerelastic}
\alias{stop_powerelastic}
\title{STOPS version of elastic scaling with powers}
\usage{
stop_powerelastic(dis, theta = c(1, 1, -2), weightmat = NULL, init = NULL,
  ndim = 2, ..., stressweight = 1, structures = c("cclusteredness",
  "clinearity", "cdependence", "cmanifoldness", "cassociation",
  "cnonmonotonicity", "cfunctionality", "ccomplexity", "cfaithfulness",
  "cregularity"), strucweight = rep(1/length(structures), length(structures)),
  strucpars, verbose = 0, type = c("additive", "multiplicative"))
}
\arguments{
\item{dis}{numeric matrix or dist object of a matrix of proximities}

\item{theta}{the theta vector of powers; the first is kappa (for the fitted distances), the second lambda (for the observed proximities) and nu as the third (fixed to -2). If a scalar for the free parameters is given it is recycled.  Defaults to 1 1 -2.}

\item{weightmat}{(optional) a matrix of nonnegative weights}

\item{init}{(optional) initial configuration}

\item{ndim}{number of dimensions of the target space}

\item{...}{additional arguments to be passed to the fitting procedure}

\item{stressweight}{weight to be used for the fit measure; defaults to 1}

\item{structures}{which strcutures to look for}

\item{strucweight}{weight to be used for the structures; defaults to 0.5}

\item{strucpars}{a list of parameters for the structuredness indices; each list element corresponds to one index in the order of the appeacrance in structures}

\item{verbose}{numeric value hat prints information on the fitting process; >2 is extremely verbose}

\item{type}{which weighting to be used in the multi-objective optimization? Either 'additive' (default) or 'multiplicative'.}
}
\value{
A list with the components
\itemize{
        \item stress: the stress
        \item stress.m: default normalized stress
        \item stoploss: the weighted loss value
        \item struc: the structuredness indices
        \item parameters: the parameters used for fitting (kappa, lambda)
        \item fit: the returned object of the fitting procedure
}
}
\description{
STOPS version of elastic scaling with powers
}
\keyword{multivariate}
