% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stops.R
\name{stop_smacofSym}
\alias{stop_smacofSym}
\title{STOPS version of smacofSym models}
\usage{
stop_smacofSym(dis, theta = c(1, 1, 1), ndim = 2, weightmat = NULL,
  init = NULL, ..., structures = c("cclusteredness", "clinearity",
  "cdependence", "cmanifoldness", "cassociation", "cnonmonotonicity",
  "cfunctionality", "ccomplexity", "cfaithfulness", "chierarchy",
  "cconvexity", "cstriatedness", "coutlying", "cskinniness", "csparsity",
  "cstringiness", "cclumpiness", "cinequality"), stressweight = 1,
  strucweight = rep(1/length(structures), length(structures)), strucpars,
  verbose = 0, type = c("additive", "multiplicative"))
}
\arguments{
\item{dis}{numeric matrix or dist object of a matrix of proximities}

\item{theta}{the theta vector for transformations; either a scalar for the lambda (proximity) transformation or a vector of at most three elements where the second is used as the free lambda parameter (the first an third would correspond to kappa and nu respecively and are internally fixed to 1)}

\item{ndim}{number of dimensions of the target space}

\item{weightmat}{(optional) a matrix of nonnegative weights}

\item{init}{(optional) initial configuration}

\item{...}{additional arguments to be passed to the fitting}

\item{structures}{which structuredness indices to be included in the loss}

\item{stressweight}{weight to be used for the fit measure; defaults to 1}

\item{strucweight}{weight to be used for the structuredness indices; ; defaults to 1/#number of structures}

\item{strucpars}{the parameters for the structuredness indices}

\item{verbose}{numeric value hat prints information on the fitting process; >2 is extremely verbose}

\item{type}{How to construct the target function for the multi objective optimization? Either 'additive' (default) or 'multiplicative'}
}
\value{
A list with the components
   \itemize{
        \item{stress:} the stress
        \item{stress.m:} default normalized stress
        \item{stoploss:} the weighted loss value
        \item{indices:} the values of the structuredness indices
        \item{parameters:} the parameters used for fitting 
        \item{fit:} the returned object of the fitting procedure
        \item{indobj:} the index objects
}
}
\description{
STOPS version of smacofSym models
}
\keyword{multivariate}
