% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stops.R
\name{stoploss}
\alias{stoploss}
\title{##High Level STOPS Function
##
##Currently only COPS, the STOPS model for c-clusteredness is implemented
##
##}
\usage{
stoploss(obj, stressweight = 1, structures = c("cclusteredness",
  "clinearity"), strucweight = rep(1/length(structures), length(structures)),
  strucpars, type = c("additive", "multiplicative"), verbose = 0)
}
\arguments{
\item{obj}{MDS object (supported are stop_sammon, stop_cmdscale, stop_smacofSym, stop_rstress, stop_powerstress, stop_smacofSphere)}

\item{stressweight}{weight to be used for the fit measure; defaults to 1}

\item{structures}{which c-structuredness indices to be included in the loss}

\item{strucweight}{the weights of the structuredness indices; defaults to 1/#number of structures}

\item{strucpars}{a list of parameters to be passed to the c-structuredness indices in the same order as the values in structures #(alternatively a named list that has the structure name as the element name)}

\item{type}{what type of weighted optimization should be used? Can be 'additive' or 'multiplicative'}

\item{verbose}{verbose output}

\item{dis}{numeric matrix or dist object of a matrix of proximities
##}

\item{structure}{what structuredness should be considered
##}

\item{...}{additional arguments to be passed to the work horses
##}
}
\value{
see \code{\link{cops}}
##
 Calculate the weighted multiobjective loss function used in STOPS
}
\description{
##High Level STOPS Function
##
##Currently only COPS, the STOPS model for c-clusteredness is implemented
##
##
}

