% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stops-package.R, R/stops.R
\docType{package}
\name{stops}
\alias{stops}
\alias{stops-package}
\title{stops: structure optimized proximity scaling}
\usage{
stops(dis, structure = "clusteredness", ...)
}
\arguments{
\item{dis}{numeric matrix or dist object of a matrix of proximities}

\item{structure}{what structuredness should be considered}

\item{...}{additional arguments to be passed to the work horses}
}
\value{
see \code{\link{cops}}
}
\description{
A package for "structure optimized proximity scaling" (STOPS), a collection of methods that fit nonlinear distance transformations in multidimensional scaling (MDS) and trade-off the fit with structure considerations to find optimal parameters. This includes the COPS method. The package contains various functions, wrappers, methods and classes for fitting, plotting and displaying different MDS models (classical scaling, Sammon mapping, elastic scaling, symmetric SMACOF, spherical SMACOF, sstress, rstress, powerStress, sammon mapping with powers and elastic scaling with powers), metaheuristics of multi-objective optimization, functions for various structuredness indices (including the OPTICS cordillera) and an interface to ELKI's OPTICS.

Currently only COPS, the STOPS model for c-clusteredness is implemented
}
\details{
The stops package provides five categories of important functions:

Models & Algorithms:
\itemize{
\item cops() and stops() ... which fit COPS and STOPS models as described in Rusch et al. (2015a) and Rusch et al. (2015b). By setting cordweight or strucweight to zero they can also be used to fit metric MDS for many different models, see below.
\item powerStressMin()... a workhorse for fitting s-stress, r-stress (de Leeuw, 2014), powerStress (e.g., Rusch et al., 2015a), Sammon mapping with power transformations (powersammon) and elastic scaling with power transformation (powerelastic). They can most conveniently be accessed via the cops or stops functions and setting stressweight=1 and cordweight or strucweight=0 or by the dedicated functions starting with cops_XXX where XXX is the method and setting stressweight=1 and cordweight=0. It uses the nested majorization algorithm for r-stress of De Leeuw(2014).
\item optics() ... An interface to ELKI's implementation of the OPTICS
}

Structuredness Indices:
\itemize{
\item cordillera() ... OPTICS cordillera
}

Optimization functions:
\itemize{
\item ljoptim() ... An (adaptive) version of the Luus-Jakola random search
}
Wrappers and convenience functions:
\itemize{
\item conf_adjust(): procrustes adjustment of configurations
\item cmdscale(), sammon(): wrappers that return S3 objects
\item cop_smacofSym(), cop_sammon(), cop_cmdscale(), cop_rstress(), cop_powerstress(),cop_smacofSphere(), cop_sammon2(), cop_elastic(), cop_sstress(), cop_powerelastic(),cop_powersammon(): cop versions of these MDS models.
\item coploss() ... a function to calculate coploss (Rusch et al., 2015a)
}

Methods:
For most of the objects returned by the high-level functions S3 classes and methods for standard generics were implemented, including print, summary, plot, plot3d, plot3dstatic.

References:
\itemize{
\item Rusch, T., Mair, P. \& Hornik, K. (2015a) COPS: Cluster optimized proximity scaling, Report 2015/1, Discussion Paper Series, Center for Empirical Research Methods, WU Vienna University of Economics and Business.
\item Rusch, T., Mair, P. \& Hornik, K. (2015b) Structuredness Indices and Augmented Nonlinear Dimension Reduction, Report 2015/X, Discussion Paper Series, Center for Empirical Research Methods, WU Vienna University of Economics and Business. \emph{forthcoming}
}

Authors: Thomas Rusch, Jan de Leeuw, Patrick Mair

Maintainer: Thomas Rusch
}
\examples{
\donttest{
data(BankingCrisesDistances)

#COPS
res<-cops(BankingCrisesDistances[,1:69])
res
summary(res)
plot(res)
plot(res,"reachplot")
plot(res,"transplot")
plot(res,"Shepard")

#OPTICS
ores<-optics(res$points,minpts=2,epsilon=100)
ores
summary(ores)
plot(ores)

#OPTICS cordillera
cres<-cordillera(res$points)
cres
summary(cres)
plot(cres)
}
\donttest{
data(BankingCrisesDistances)
res1<-stops(BankingCrisesDistances[,1:69],structure="clusteredness",loss="strain",verbose=0)
res1
}
}
\keyword{clustering}
\keyword{multivariate}

